+++ packages/manager/modules/pci/src/projects/project/sidebar/sidebar.controller.js
@@ -1,12 +1,5 @@
-import clone from 'lodash/clone';
-import filter from 'lodash/filter';
-import forEach from 'lodash/forEach';
-import get from 'lodash/get';
-import has from 'lodash/has';
 import isEmpty from 'lodash/isEmpty';
-import map from 'lodash/map';
-import mapValues from 'lodash/mapValues';
-import sortBy from 'lodash/sortBy';
+import isNil from 'lodash/isNil';
 
 import { MENU } from './sidebar.constant';
 
@@ -18,20 +11,20 @@ export default class SidebarController {
     $rootScope,
     $stateParams,
     $transitions,
+    atInternet,
     coreConfig,
     OvhApiServices,
     OvhApiCloudProject,
-    SidebarMenu,
   ) {
     this.$translate = $translate;
     this.$transitions = $transitions;
     this.$rootScope = $rootScope;
 
     this.$stateParams = $stateParams;
+    this.atInternet = atInternet;
     this.coreConfig = coreConfig;
     this.OvhApiServices = OvhApiServices;
     this.OvhApiCloudProject = OvhApiCloudProject;
-    this.SidebarMenu = SidebarMenu;
 
     this.isOpen = false;
     this.isLoading = false;
@@ -49,71 +42,23 @@ export default class SidebarController {
 
     this.isLoading = true;
 
-    return this.SidebarMenu.setInitializationPromise(
-      this.OvhApiCloudProject
-        .v6()
-        .get({
-          serviceName,
-        })
-        .$promise
-        .then((project) => {
-          this.project = project;
-
-          return this
-            .$translate
-            .refresh() // Waiting for translations ./translations
-            .then(() => new Promise((resolve) => {
-              this.SidebarMenu.setShouldToggleMenuItemOpenState(false);
-              this.SidebarMenu.clearMenuItems();
-
-              forEach(MENU.filter(el => el !== null), ({
-                options,
-                regions,
-                subitems,
-                translation,
-              }) => {
-                if (
-                  (Array.isArray(regions) && this.coreConfig.isRegion(regions))
-                  || !Array.isArray(regions)
-                ) {
-                  const menuItem = this.SidebarMenu.addMenuItem(Object.assign({
-                    title: this.$translate.instant(translation),
-                    allowSubItems: true,
-                    isOpen: true,
-                    stateParams: {
-                      projectId: this.project.project_id,
-                    },
-                  }, options));
-
-                  /* eslint-disable no-shadow */
-                  forEach(subitems.filter(el => el !== null), ({
-                    options,
-                    regions,
-                    translation,
-                  }) => {
-                    /* eslint-enable no-shadow */
-                    if (
-                      (Array.isArray(regions) && this.coreConfig.isRegion(regions))
-                      || !Array.isArray(regions)
-                    ) {
-                      this.SidebarMenu.addMenuItem(Object.assign({
-                        title: this.$translate.instant(translation),
-                      }, options), menuItem);
-                    }
-                  });
-                }
-              });
-              resolve();
-            }));
-        })
-        .finally(() => {
-          this.$rootScope.$broadcast('sidebar:loaded');
-          this.isLoading = false;
-        }),
-    );
+    return this.OvhApiCloudProject
+      .v6()
+      .get({
+        serviceName,
+      })
+      .$promise
+      .then((project) => {
+        this.project = project;
+      })
+      .finally(() => {
+        this.$rootScope.$broadcast('sidebar:loaded');
+        this.isLoading = false;
+      });
   }
 
   $onInit() {
+    this.MENU = MENU.filter(({ regions }) => isNil(regions) || this.coreConfig.isRegion(regions));
     let currentProjectId = this.$stateParams.projectId;
     this.setProject(this.$stateParams.projectId);
     this.$transitions.onSuccess({}, () => {
@@ -125,92 +70,10 @@ export default class SidebarController {
     });
   }
 
-  buildMenu(services) {
-    const menuItems = this.buildMenuTreeConfig(services);
-    this.buildMenuItems(menuItems);
-    this.$rootScope.$broadcast('sidebar:loaded');
-  }
-
-  hasMenuItemConfig(serviceName) {
-    return has(this.SIDEBAR_CONFIG, serviceName);
-  }
-
-  getMenuItemConfig(serviceName) {
-    const config = clone(get(this.SIDEBAR_CONFIG, serviceName, this.SIDEBAR_CONFIG.default));
-
-    config.id = serviceName;
-    if (!has(config, 'title')) {
-      config.title = serviceName;
-    }
-    config.title = this.$translate.instant(config.title);
-
-    if (has(config, 'error')) {
-      config.error = this.$translate.instant(config.error);
-    }
-    config.allowSearch = false;
-    config.allowSubItems = true;
-    return config;
-  }
-
-  getServiceMenuItemConfig(service) {
-    const serviceConfig = get(
-      this.SIDEBAR_STATE_MAPPING_SERVICE,
-      service.route.path,
-      this.SIDEBAR_STATE_MAPPING_SERVICE.default,
-    );
-    return {
-      id: service.resource.name,
-      title: service.resource.displayName || service.resource.name,
-      prefix: this.$translate.instant(serviceConfig.prefix),
-      state: serviceConfig.state,
-      stateParams: mapValues(serviceConfig.stateParams, param => get(service, param, '')),
-    };
-  }
-
-  buildMenuTreeConfig(services) {
-    const filteredServices = filter(services, service => this.hasMenuItemConfig(service.name));
-    return sortBy(
-      map(filteredServices, (service) => {
-        const menuItem = this.getMenuItemConfig(service.name);
-        if (has(service, 'subServices') && !isEmpty(service.subServices)) {
-          menuItem.subServices = this.buildMenuTreeConfig(service.subServices);
-        }
-        return menuItem;
-      }),
-      'title',
-    );
-  }
-
-  buildMenuItems(menuTree, parent) {
-    forEach(menuTree, (menuItemConfig) => {
-      let children = [];
-      if (has(menuItemConfig, 'subServices')) {
-        children = menuItemConfig.subServices;
-      }
-      let menuItem;
-
-      if (parent) {
-        menuItem = this.SidebarMenu.addMenuItem(menuItemConfig, parent);
-      } else {
-        menuItem = this.SidebarMenu.addMenuItem(menuItemConfig);
-      }
-
-      if (!isEmpty(children)) {
-        this.buildMenuItems(children, menuItem);
-      } else {
-        menuItem.onLoad = () => this.loadServices(menuItemConfig.id);
-      }
+  onMenuItemClick({ id }) {
+    this.atInternet.trackClick({
+      name: `public-cloud_menu_${id}`,
+      type: 'action',
     });
-  }
-
-  loadServices(serviceName) {
-    return this.OvhApiServices.Aapi().get({ service: serviceName }).$promise
-      .then((services) => {
-        const parent = this.SidebarMenu.getItemById(serviceName);
-
-        forEach(services, (service) => {
-          this.SidebarMenu.addMenuItem(this.getServiceMenuItemConfig(service), parent);
-        });
-      });
   }
 }